all tests restructured
