+++ src/app.rs
@@ -532,17 +532,16 @@ impl App {
 //   ██║   ███████╗███████║   ██║   ███████║
 //   ╚═╝   ╚══════╝╚══════╝   ╚═╝   ╚══════╝
 
-
+#[cfg(test)]
 mod unit_app_tests {
     use crate::app::*;
 
     fn create_app() -> App {
-        let mut app = App::new();
+        let app = App::new();
         app
     }
 
-    
-    
+
     #[test]
     fn test_toggle_to_command_line() {
         let mut app = create_app();
@@ -572,9 +571,9 @@ mod unit_app_tests {
     }
     
 }
-
+#[cfg(test)]
 mod unit_app_command_line_tests {
-    use calliglyph::app::App;
+    use crate::app::App;
 
     
 
@@ -627,5 +626,296 @@ mod unit_app_command_line_tests {
     }
     
 }
+#[cfg(test)]
+mod unit_app_cutcopy_tests{
+    use crate::cursor::CursorPosition;
+    use crate::app::*;
+
+    fn create_app_with_editor_content(vec: Vec<String>) -> App {
+        let mut app = App::new();
+        app.editor.editor_content = vec;
+        app.editor.editor_height = 10; //since testing doesnt start ui.rs, height isnt set
+        app
+    }
 
 
+    //copy selected text
+    #[test]
+    fn test_copy_single_line_selection() {
+        let mut app = create_app_with_editor_content(vec!["Hello, world!".to_string()]);
+        app.editor.text_selection_start = Some(CursorPosition { x: 7, y: 0 });
+        app.editor.text_selection_end = Some(CursorPosition { x: 12, y: 0 });
+
+        let result = app.copy_selected_text();
+
+        assert!(result.is_ok());
+        assert_eq!(app.clipboard.copied_text, vec!["world".to_string()]);
+    }
+
+    #[test]
+    fn test_copy_multi_line_selection() {
+        let mut app = create_app_with_editor_content(vec![
+            "Hello,".to_string(),
+            " world!".to_string(),
+            " Rust".to_string(),
+        ]);
+        app.editor.text_selection_start = Some(CursorPosition { x: 4, y: 0 });
+        app.editor.text_selection_end = Some(CursorPosition { x: 3, y: 2 });
+
+        let result = app.copy_selected_text();
+
+        assert!(result.is_ok());
+        assert_eq!(
+            app.clipboard.copied_text,
+            vec!["o,", " world!", " Ru"]
+                .into_iter()
+                .map(String::from)
+                .collect::<Vec<String>>()
+        );
+    }
+
+    #[test]
+    fn test_copy_no_selection() {
+        let mut app = create_app_with_editor_content(vec!["Hello, world!".to_string()]);
+        app.editor.text_selection_start = None;
+        app.editor.text_selection_end = None;
+
+        let result = app.copy_selected_text();
+
+        assert!(result.is_err());
+        assert!(app.clipboard.copied_text.is_empty());
+    }
+
+    //cut selected text
+    #[test]
+    fn test_cut_single_line_selection() {
+        let mut app = create_app_with_editor_content(vec!["Hello, world!".to_string()]);
+        app.editor.text_selection_start = Some(CursorPosition { x: 7, y: 0 });
+        app.editor.text_selection_end = Some(CursorPosition { x: 12, y: 0 });
+
+        let result = app.cut_selected_text();
+
+        assert!(result.is_ok());
+        assert_eq!(app.clipboard.copied_text, vec!["world".to_string()]);
+        assert!(app.editor.text_selection_start.is_none());
+        assert!(app.editor.text_selection_end.is_none());
+    }
+
+    #[test]
+    fn test_cut_multi_line_selection() {
+        let mut app = create_app_with_editor_content(vec![
+            "Hello,".to_string(),
+            " world!".to_string(),
+            " Rust".to_string(),
+        ]);
+        app.editor.text_selection_start = Some(CursorPosition { x: 4, y: 0 });
+        app.editor.text_selection_end = Some(CursorPosition { x: 3, y: 2 });
+
+        let result = app.cut_selected_text();
+
+        assert!(result.is_ok());
+        assert_eq!(
+            app.clipboard.copied_text,
+            vec!["o,", " world!", " Ru"]
+                .into_iter()
+                .map(String::from)
+                .collect::<Vec<String>>()
+        );
+        assert!(app.editor.text_selection_start.is_none());
+        assert!(app.editor.text_selection_end.is_none());
+    }
+
+    #[test]
+    fn test_cut_no_selection() {
+        let mut app = create_app_with_editor_content(vec!["Hello, world!".to_string()]);
+        app.editor.text_selection_start = None;
+        app.editor.text_selection_end = None;
+
+        let result = app.cut_selected_text();
+
+        assert!(result.is_err());
+        assert!(app.clipboard.copied_text.is_empty());
+        assert!(app.editor.text_selection_start.is_none());
+        assert!(app.editor.text_selection_end.is_none());
+    }
+
+    //paste selected text
+    #[test]
+    fn test_paste_single_line() {
+        let mut app = create_app_with_editor_content(vec![
+            "Hello, world!".to_string(),
+            "This is a test.".to_string(),
+            "Another line.".to_string(),
+        ]);
+        app.clipboard.copy(&vec!["PASTED".to_string()]);
+        app.editor.cursor.x = 8;
+        app.editor.cursor.y = 0;
+
+        app.paste_selected_text().unwrap();
+
+        assert_eq!(
+            app.editor.editor_content,
+            vec![
+                "Hello, wPASTEDorld!".to_string(),
+                "This is a test.".to_string(),
+                "Another line.".to_string(),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_paste_multiline() {
+        let mut app = create_app_with_editor_content(vec![
+            "Hello, world!".to_string(),
+            "This is a test.".to_string(),
+            "Another line.".to_string(),
+        ]);
+        app.clipboard
+            .copy(&vec!["First".to_string(), "Second ".to_string()]);
+        app.editor.cursor.x = 5;
+        app.editor.cursor.y = 1;
+
+        app.paste_selected_text().unwrap();
+
+        assert_eq!(
+            app.editor.editor_content,
+            vec![
+                "Hello, world!".to_string(),
+                "This First".to_string(),
+                "Second is a test.".to_string(),
+                "Another line.".to_string(),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_paste_single_line_special_characters() {
+        let mut app = create_app_with_editor_content(vec![
+            "Hello, wᚠᚠᚠᚠorld!".to_string(),
+            "This is a test.".to_string(),
+            "Another line.".to_string(),
+        ]);
+        app.clipboard.copy(&vec!["PASTED".to_string()]);
+        app.editor.cursor.x = 10;
+        app.editor.cursor.y = 0;
+
+        app.paste_selected_text().unwrap();
+
+        assert_eq!(
+            app.editor.editor_content,
+            vec![
+                "Hello, wᚠᚠPASTEDᚠᚠorld!".to_string(),
+                "This is a test.".to_string(),
+                "Another line.".to_string(),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_paste_multiline_special_charaters() {
+        let mut app = create_app_with_editor_content(vec![
+            "Hello, world!".to_string(),
+            "This ᚠᚠᚠᚠis a test.".to_string(),
+            "Another line.".to_string(),
+        ]);
+        app.clipboard
+            .copy(&vec!["First".to_string(), "Second ".to_string()]);
+        app.editor.cursor.x = 7;
+        app.editor.cursor.y = 1;
+
+        app.paste_selected_text().unwrap();
+
+        assert_eq!(
+            app.editor.editor_content,
+            vec![
+                "Hello, world!".to_string(),
+                "This ᚠᚠFirst".to_string(),
+                "Second ᚠᚠis a test.".to_string(),
+                "Another line.".to_string(),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_paste_at_start_of_line() {
+        let mut app = create_app_with_editor_content(vec![
+            "Hello, world!".to_string(),
+            "This is a test.".to_string(),
+            "Another line.".to_string(),
+        ]);
+        app.clipboard.copy(&vec!["NewStart".to_string()]);
+        app.editor.cursor.x = 0;
+        app.editor.cursor.y = 2;
+
+        app.paste_selected_text().unwrap();
+
+        assert_eq!(
+            app.editor.editor_content,
+            vec![
+                "Hello, world!".to_string(),
+                "This is a test.".to_string(),
+                "NewStartAnother line.".to_string(),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_paste_at_end_of_line() {
+        let mut app = create_app_with_editor_content(vec![
+            "Hello, world!".to_string(),
+            "This is a test.".to_string(),
+            "Another line.".to_string(),
+        ]);
+        app.clipboard.copy(&vec!["END".to_string()]);
+        app.editor.cursor.x = 13;
+        app.editor.cursor.y = 0;
+
+        app.paste_selected_text().unwrap();
+
+        assert_eq!(
+            app.editor.editor_content,
+            vec![
+                "Hello, world!END".to_string(),
+                "This is a test.".to_string(),
+                "Another line.".to_string(),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_paste_with_empty_copied_text() {
+        let mut app = create_app_with_editor_content(vec![
+            "Hello, world!".to_string(),
+            "This is a test.".to_string(),
+            "Another line.".to_string(),
+        ]);
+        app.clipboard.copy(&vec![]);
+        app.editor.cursor.x = 5;
+        app.editor.cursor.y = 1;
+
+
+        assert!(app.paste_selected_text().is_err());
+        assert_eq!(
+            app.editor.editor_content,
+            vec![
+                "Hello, world!".to_string(),
+                "This is a test.".to_string(),
+                "Another line.".to_string(),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_paste_into_empty_editor() {
+        let mut app = create_app_with_editor_content(vec![]);
+        app.clipboard
+            .copy(&vec!["Hello".to_string(), "World".to_string()]);
+
+        app.paste_selected_text().unwrap();
+
+        assert_eq!(
+            app.editor.editor_content,
+            vec!["Hello".to_string(), "World".to_string()]
+        );
+    }
+}
